Project_dir/requirements.txt
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==19.2.0
arrow==0.15.6
asgiref==3.3.1
astroid==2.4.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
black==19.10b0
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
collection==0.1.6
colorama==0.4.3
comtypes==1.1.9
constantly==15.1.0
cookiecutter==1.7.2
coverage==5.1
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.41.0
django-coverage-plugin==1.8.0
django-crispy==1.8.1
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-decouple==2.0
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.4.0
django-import-export==2.1.0
django-model-utils==4.0.0
django-moneyfield==0.2.1
django-notifications-hq==1.6.0
django-redis==4.11.0
django-rest-auth==0.9.5
django-seed==0.2.2
django-stubs==1.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
djangorestframework-datatables==0.5.2
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
docutils==0.16
drf-renderer-xlsx==0.3.7
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==5.8.0
filelock==3.0.12
flake8==3.8.1
flake8-isort==3.0.0
future==0.18.2
googletrans==3.0.0
greenlet==1.1.2
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hiredis==1.0.1
hpack==3.0.0
hstspreload==2020.12.22
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
huepy==1.2.1
hyperframe==5.2.0
hyperlink==19.0.0
identify==1.4.16
idna==2.9
imageio==2.12.0
imageio-ffmpeg==0.4.5
imagesize==1.2.0
importlib-metadata==1.6.0
importlib-resources==1.5.0
incremental==17.5.0
inflect==4.1.0
instabot==0.117.0
install==1.3.3
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jinja2-time==0.2.0
jsonfield==3.1.0
lazy-object-proxy==1.4.3
lxml==4.5.2
Markdown==3.2.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.3
more-itertools==8.3.0
MouseInfo==0.1.3
moviepy==1.0.3
msgpack==0.6.2
mypy==0.770
mypy-extensions==0.4.3
mysql-connector-python==8.0.22
mysqlclient==2.0.3
nodeenv==1.3.5
num2words==0.5.10
numpy==1.19.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.4
pandas==1.1.2
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.4.0
proglog==0.1.9
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyGetWindow==0.0.9
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==2.0.1
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
PyMsgBox==1.0.9
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.1
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
pytest==5.4.2
pytest-django==3.9.0
pytest-sugar==0.9.3
python-dateutil==2.8.1
python-docx==0.8.10
python-slugify==4.0.0
python3-openid==3.1.0
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.4
pywin32==300
pywinauto==0.6.8
PyYAML==5.3.1
redis==3.5.0
regex==2020.5.14
reportlab==3.5.44
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.16.0
rfc3986==1.4.0
schedule==1.1.0
service-identity==18.1.0
six==1.15.0
sniffio==1.2.0
snowballstemmer==2.0.0
soupsieve==2.1
SpeechRecognition==3.8.1
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.28
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==2.0.0
termcolor==1.1.0
testfixtures==6.14.1
text-unidecode==1.3
toml==0.10.1
tqdm==4.62.3
traitlets==4.3.3
twilio==6.47.0
Twisted @ file:///C:/Users/asifk/Downloads/Twisted-20.3.0-cp38-cp38-win_amd64.whl
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
unicodecsv==0.14.1
Unipath==1.1
urllib3==1.25.11
virtualenv==20.0.21
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.0.1
word2number==1.1
wrapt==1.12.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.1.0
